[refactor] DoseFooter 做成獨立元件/ 加入comment 按鈕
