+++ src/views/Pen.vue
@@ -16,7 +16,8 @@
   import { useWorkStore } from '@/stores/useWorkStore';
   import { useAuthStore } from '@/stores/useAuthStore';
   import { useHandleSave } from '@/utils/handleWorkSave';
-  import { useRoute, useRouter } from 'vue-router'
+  import { useRoute, useRouter } from 'vue-router';
+  import DoseFooter from '@/components/Editor/DoseFooter.vue';
 
   const route = useRoute();
   const router = useRouter();
@@ -61,7 +62,6 @@
   onBeforeUnmount(() => {
     window.removeEventListener('keydown', handleKeydown, true)
   })
-
 
   const penHeader = ref(null)
   const htmlCode = ref('');
@@ -121,6 +121,7 @@
       currentWork.value.cdns,
       currentWork.value.links,
       currentWork.value.viewMode,
+      currentWork.value.tabSize,
       currentWork.value.isAutoSave,
       currentWork.value.isAutoPreview,
       currentWork.value.isPrivate,
@@ -386,28 +387,6 @@
   const handleRunPreview = () => {
     previewRef.value?.runPreview()
   }
-
-  const handleMoveToTrash = async () => {
-    const confirmed = window.confirm('確定要將這個作品移至垃圾桶嗎？此操作可以在垃圾桶中還原。');
-    if (!confirmed) return;
-
-  try {
-    const id = currentId.value;
-    const success = await moveToTrash(id);
-
-    if (success) {
-      console.log(`作品 ID: ${id} 已丟入垃圾桶`);
-      await router.push({ path: '/your-work' });
-    } else {
-      console.warn(`移動失敗：伺服器未回傳成功狀態`);
-      alert('無法丟入垃圾桶，請稍後再試');
-    }
-  } catch (error) {
-    console.error('丟入垃圾桶失敗：', error);
-    alert('無法丟入垃圾桶');
-  }
-};
-
 </script>
 
 <template>
@@ -613,14 +592,6 @@
         
       </div>
     </main>
-
-    <footer class="h-8 w-full flex relative justify-between items-center py-[.2rem] px-3 bg-cc-14 text-white">
-        <div class="flex items-center h-full">
-          <EditorSmallButton class="hover:bg-cc-12" @buttonClick="toggleConsole">Console</EditorSmallButton>
-        </div>
-        <div class="flex items-center h-full">
-          <EditorSmallButton class="hover:bg-cc-red" @click.prevent="handleMoveToTrash">Delete</EditorSmallButton>
-        </div>
-    </footer>
+    <DoseFooter @toggle-console="toggleConsole"/>
   </div>
 </template>
